+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Generation error when using `poketch_apps = false`.
 * Issues running data generation with Python version < 3.12.
 * Hardlock when obtaining the Pokétch from a local location.
+* Random text frame is now random.
+# Changed
+* Game options now use the default value if there is no corresponding entry in the YAML.
+* Game options now raise exceptions if invalid values are entered.
 
 ## [0.0.1] - 2025-08-26
 The first release of this project.